+++ R/TAF-package.R
@@ -90,6 +90,7 @@
 #' }
 #' \emph{Administrative tools, rarely used in scripts:}
 #' \tabular{ll}{
+#'   \code{\link{check.software}}  \tab check global package versions\cr
 #'   \code{\link{clean.data}}      \tab clean boot data\cr
 #'   \code{\link{clean.library}}   \tab clean TAF library\cr
 #'   \code{\link{clean.software}}  \tab clean TAF software\cr